feat(auth): caller-role provider hardening + read-path descent tracing
Caller-role provider strand: ICallerPositionAccessor removed (position now
resolved inside the caller-role providers), AuthorizeAppService and the
instance query/retry services take ICallerRoleResolver, MorphIdm resolver
and its DI extensions reworked, current-user header extensions and forward
headers updated accordingly; tests moved to CallerRoleHeaderExtensionsTests.

Read-path tracing strand: Auth.ResolveRoles span (AuthorizationActivityHelper)
and Subflow.Descend/{flow} descent ladder (InstanceReadActivityHelper +
SubflowDescentContext, X-Subflow-Depth header) with tests; trace-span-tree
doc and host appsettings updated for the new sources.

Build plumbing: AetherPackageVersion 1.0.40-local via repo-local aether feed
(nuget.config) — must be set back to a published version before this branch
merges to master (documented inline).

+++ src/BBT.Workflow.Domain/Logging/WorkflowLogs.cs
@@ -3265,7 +3265,7 @@ public static partial class WorkflowLogs
         this ILogger logger,
         string provider,
         int roleCount,
-        long elapsedMs);
+        double elapsedMs);
 
     /// <summary>
     /// Logs when the provider answered that the caller has no operation set at all. This is a valid
